refactor(update_imports): isolate CLI and migrate tests
